migrate tests to robolectric
the ones that fail need some heavy mocking
also notice: they run in seconds, not minutes
