+++ app/build.gradle
@@ -74,6 +74,12 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     testImplementation 'androidx.test.ext:junit:1.1.3'
     testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.0'
+    testImplementation 'org.robolectric:robolectric:4.6' // add all test dependencies
+    testImplementation 'androidx.test:core-ktx:1.4.0' // necessary for kotlin tests
+    testImplementation 'androidx.test:runner:1.4.0'
+    testImplementation 'androidx.test:rules:1.4.0'
+    testImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    testImplementation 'androidx.test.espresso:espresso-intents:3.4.0'
     androidTestImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.0'
     androidTestImplementation 'androidx.test:core-ktx:1.4.0' // necessary for kotlin tests
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
