+++ app/src/test/java/com/github/sdpsharelook/MainActivityTest.kt
@@ -19,15 +19,15 @@ class MainActivityTest {
 
     @Test
     fun testMainActivity() {
-//        Intents.init()
-//        val result =  Intent()
-//        Espresso.onView(ViewMatchers.withId(R.id.edit_text_name))
-//            .perform(ViewActions.typeText("World"))
-//        Espresso.onView(ViewMatchers.withId(R.id.button_greet))
-//            .perform(ViewActions.click())
-//        Intents.intended(IntentMatchers.toPackage("com.github.sdpsharelook"))
-//        Intents.intended(IntentMatchers.hasExtraWithKey(EXTRA_MESSAGE))
-//        Intents.intended(IntentMatchers.hasExtra(EXTRA_MESSAGE, "World"))
-//        Intents.release()
+        Intents.init()
+        Intent()
+        Espresso.onView(ViewMatchers.withId(R.id.edit_text_name))
+            .perform(ViewActions.typeText("World"))
+        Espresso.onView(ViewMatchers.withId(R.id.button_greet))
+            .perform(ViewActions.click())
+        Intents.intended(IntentMatchers.toPackage("com.github.sdpsharelook"))
+        Intents.intended(IntentMatchers.hasExtraWithKey(EXTRA_MESSAGE))
+        Intents.intended(IntentMatchers.hasExtra(EXTRA_MESSAGE, "World"))
+        Intents.release()
     }
 }
