+++ app/src/test/java/com/github/sdpsharelook/SectionActivityTest.kt
@@ -3,18 +3,17 @@ package com.github.sdpsharelook
 
 import android.view.View
 import android.view.ViewGroup
-import androidx.test.espresso.Espresso.onData
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.*
-import androidx.test.espresso.assertion.ViewAssertions.*
+import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
+import androidx.test.ext.junit.rules.ActivityScenarioRule
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
-import androidx.test.rule.ActivityTestRule
-import androidx.test.runner.AndroidJUnit4
 import com.github.sdpsharelook.Section.SectionActivity
 import org.hamcrest.Description
 import org.hamcrest.Matcher
-import org.hamcrest.Matchers.*
+import org.hamcrest.Matchers.allOf
 import org.hamcrest.TypeSafeMatcher
 import org.junit.Rule
 import org.junit.Test
@@ -26,7 +25,7 @@ class SectionActivityTest {
 
     @Rule
     @JvmField
-    var mActivityTestRule = ActivityTestRule(SectionActivity::class.java)
+    var mActivityTestRule = ActivityScenarioRule(SectionActivity::class.java)
 
     @Test
     fun sectionActivityTest() {
