+++ app/src/test/java/com/github/sdpsharelook/TranslateActivityTest.kt
