+++ app/src/test/java/com/github/sdpsharelook/translate/TranslatorTest.kt
