+++ app/src/test/resources/robolectric.properties
@@ -1,0 +1,4 @@
+sdk=30
+instrumentedPackages=androidx.loader.content
+#application=dagger.hilt.android.testing.HiltTestApplication
+qualifiers=w360dp-h640dp-hdpi
